pyproject.toml
[tool.poetry]
name = "pros-interface"
version = "0.1.0"
description = ""
authors = ["Richard Hadden <richard.hadden@oeaw.ac.at>"]
readme = "README.md"
packages = [{include = "pros_interface"}]

[tool.poetry.dependencies]
python = "^3.11"
pynpm = "^0.1.2"
camel-converter = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pytest-playwright = "^0.3.3"
pytest-asyncio = "^0.21.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "../neomodel/",
  "../pros-core",
  "../pros-interface"
]
